chore(spec): externalize wasm runtime tests
